Text System implemented
when approaching the NPC you can now press E (as displayed on screen) and the camera will focus and zoom in. Text will then appear on the screen letter by letter. (there is currently no way to exit out of the text so I will implement that after its connected to the xml